// src/utils/server/backend/model/nutrient-value.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface NutrientValue
 */
export interface NutrientValue {
  /**
   *
   * @type {string}
   * @memberof NutrientValue
   */
  nutrient: string;
  /**
   *
   * @type {number}
   * @memberof NutrientValue
   */
  value?: number | null;
  /**
   *
   * @type {string}
   * @memberof NutrientValue
   */
  unit?: string | null;
}